# labels: improve Get method for stringlabels build