Inline one call to `decodeString`, and skip decoding the value string until we find a match for the name.

Check the first character before doing full string comparison, and exit early if that character is greater than the target name, since labels are held in order.

Also, make sure the tests work with varying lengths - it's not typical for Prometheus labels to all be the same length.

Benchmarks:
```
name                                             old time/op  new time/op  delta
Labels_Get/with_5_labels/get_first_label-8       5.32ns ± 0%  5.33ns ± 1%     ~     (p=0.368 n=6+6)
Labels_Get/with_5_labels/get_middle_label-8      10.4ns ± 0%   8.4ns ± 3%  -18.42%  (p=0.002 n=6+6)
Labels_Get/with_5_labels/get_last_label-8        15.5ns ± 0%  10.9ns ± 0%  -29.51%  (p=0.002 n=6+6)
Labels_Get/with_5_labels/get_not-found_label-8   15.6ns ± 0%   5.4ns ± 0%  -65.53%  (p=0.004 n=5+6)
Labels_Get/with_10_labels/get_first_label-8      5.31ns ± 0%  5.31ns ± 0%     ~     (p=0.403 n=5+6)
Labels_Get/with_10_labels/get_middle_label-8     20.4ns ± 0%  12.7ns ± 1%  -37.58%  (p=0.004 n=5+6)
Labels_Get/with_10_labels/get_last_label-8       30.5ns ± 2%  19.9ns ± 0%  -34.53%  (p=0.002 n=6+6)
Labels_Get/with_10_labels/get_not-found_label-8  30.7ns ± 3%   5.4ns ± 0%  -82.45%  (p=0.004 n=6+5)
Labels_Get/with_30_labels/get_first_label-8      5.30ns ± 0%  5.45ns ± 2%   +2.82%  (p=0.004 n=5+6)
Labels_Get/with_30_labels/get_middle_label-8     49.5ns ± 0%  36.8ns ± 0%  -25.73%  (p=0.004 n=6+5)
Labels_Get/with_30_labels/get_last_label-8        109ns ± 0%    78ns ± 0%  -29.02%  (p=0.004 n=6+5)
Labels_Get/with_30_labels/get_not-found_label-8   105ns ± 1%     5ns ± 1%  -94.86%  (p=0.002 n=6+6)
```

[I feel it should be possible to go a lot faster, but I spent several hours trying different ideas and nothing was better than this.]